feat: implement main process bootstrap and update IPC handlers for application state management
